requirements/dev.txt
bootstrap-admin==0.3.7.1
Django==1.10.2
django-appconf==1.0.2
django-bower==5.2.0
django-ckeditor==5.1.1
django-compressor==2.1
rcssmin==1.0.6
rjsmin==1.0.12
six==1.10.0
